Discourse: enable mirror schema upgrading
There's a small change I have in mind to upgrade to a v7 database.
However it doesn't warrant throwing out your existing cache, so I've
created a way to upgrade versions.

For this PR I've dug up the v5 (previous) schema and added an upgrade
to the current v6 as an example. This so the upgrading PR can be reviewed
separately from the v7 schema I'll suggest in a follow-up.