Discourse: use plaintext cache keys
In a similar vain to #1685, make the cache keys human readable.

Previously we had base64 and no prefix, which is case-sensitive.
So we were not scoping to our plugin, nor supporting case-insensitive
filesystems.

The fields preserved from a URL are the protocol, hostname and optional
custom port. Using `_` as the separator for those.